+++ jest.config.js
@@ -25,7 +25,7 @@ module.exports = {
     ],
     coverageThreshold: {
       global: {
-        branches: 74,
+        branches: 75,
         functions: 80,
         lines: 80,
         statements: 80,